[Backend] Fixed some types and moved some files to make the most sense
