+++ Backend/src/types/index.ts
@@ -1,2 +1,4 @@
-export {APIResponse, OfficerInfoAPIResponse} from "./ResponseTypes";
-export {DefaultReturn} from "./ServiceReturnTypes";
+export {ConfigTypes, StaticConfigTypes} from "./config-types";
+
+export {APIResponse, OfficerInfoAPIResponse} from "./response-types";
+export {DefaultReturn} from "./service-return-types";
